requirements.txt
Django==5.2.8
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
drf-spectacular==0.27.0
mysqlclient==2.2.7
python-decouple==3.8
PyJWT==2.10.1
